+++ packages/create-widget/template/server/router/playground/index.js
@@ -1,0 +1,3 @@
+const playground = require('./playground');
+
+module.exports = playground;