+++ packages/create-widget/template/server/server.js
@@ -7,6 +7,14 @@ const { app } = require('./app');
 
 const serverConfig = config.get('server');
 
+process.on('uncaughtException', (error) => {
+  console.error(error);
+});
+
+process.on('unhandledRejection', (error) => {
+  console.error(error);
+});
+
 if (!serverConfig.clusters || !cluster.isMaster) {
   app.listen(config.get('server.port'), () => {
     console.log(`listen on localhost:${config.get('server.port')}`); // eslint-disable-line no-console