+++ packages/create-widget/views/hyper/template/src/component/View.js
@@ -1,8 +1,13 @@
 import Welcome from './Welcome';
 import Counter from './Counter';
 import WidgetDescription from './WidgetDescription';
+import ErrorView from './ErrorView';
 
 export default function View(widget, render) {
+  if (widget.error && widget.error.status) {
+    return render`${ErrorView(widget)}`;
+  }
+
   return render`
       <div class='merkur__page'>
         <div class='merkur__headline'>