+++ packages/create-widget/views/preact/template/src/component/ErrorView.jsx
@@ -1,0 +1,11 @@
+import { h } from 'preact'; // eslint-disable-line no-unused-vars
+
+export default function ErrorView({ error }) {
+  return (
+    <div className="merkur__error">
+      <h1>Status: {error.status}</h1>
+      <h2>Message: {error.message}</h2>
+      <pre>{error.stack}</pre>
+    </div>
+  );
+}