+++ packages/create-widget/views/uhtml/template/src/component/View.js
@@ -1,8 +1,13 @@
 import Welcome from './Welcome';
 import Counter from './Counter';
 import WidgetDescription from './WidgetDescription';
+import ErrorView from './ErrorView';
 
 export default function View(widget) {
+  if (widget.error && widget.error.status) {
+    return ErrorView(widget);
+  }
+
   return widget.$dependencies.html`
       <div class='merkur__page'>
         <div class='merkur__headline'>