+++ packages/integration/src/index.js
@@ -116,7 +116,7 @@ function loadScriptAssets(assets) {
 }
 
 function loadAssets(assets) {
-  return Promise.all(loadScriptAssets(assets), loadStyleAssets(assets));
+  return Promise.all([loadScriptAssets(assets), loadStyleAssets(assets)]);
 }
 
 export { testScript, loadAssets, loadStyleAssets, loadScriptAssets };